+++ tests/test_app/TestApp/Plugin/TestPlugin/src/View/Cell/PluginExampleCell.php
@@ -1,0 +1,11 @@
+<?php
+namespace TestPlugin\View\Cell;
+
+use Cake\View\Cell;
+
+class PluginExampleCell extends Cell
+{
+    public function display()
+    {
+    }
+}